+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Permissive deserialization of Collection temporal extents ([#1222](https://github.com/stac-utils/pystac/pull/1222))
+
 ### Fixed
 
 - Update usage of jsonschema ([#1215](https://github.com/stac-utils/pystac/pull/1215))
@@ -9,7 +13,6 @@
 ### Deprecated
 
 - `pystac.validation.local_validator.LocalValidator` ([#1215](https://github.com/stac-utils/pystac/pull/1215))
-
 
 ## [v1.8.3] - 2023-07-12